Refactor includes: replace common_util with specific utility headers and add StringViewStreamBuf
