+++ src/common/util/src/os/win/win_mmap_object.cpp
@@ -11,8 +11,8 @@
 #include <thread>
 #include <vector>
 
-#include "openvino/util/hash_util.hpp"
 #include "openvino/util/file_util.hpp"
+#include "openvino/util/hash_util.hpp"
 #include "openvino/util/memory.hpp"
 #include "openvino/util/mmap_object.hpp"
