+++ src/core/src/shape.cpp
@@ -7,6 +7,7 @@
 #include "openvino/core/shape_util.hpp"
 #include "openvino/core/validation_util.hpp"
 #include "openvino/util/common_util.hpp"
+#include "openvino/util/string_view_streambuf.hpp"
 
 std::ostream& ov::operator<<(std::ostream& s, const Shape& shape) {
     return s << "[" << ov::util::join<std::ostream>(shape, ",") << "]";
