+++ src/core/xml_util/include/openvino/xml_util/xml_deserialize_util.hpp
@@ -18,6 +18,7 @@
 #include "openvino/opsets/opset.hpp"
 #include "openvino/runtime/aligned_buffer.hpp"
 #include "openvino/util/common_util.hpp"
+#include "openvino/util/string_view_streambuf.hpp"
 
 namespace ov::util {
 struct GenericLayerParams;
