+++ src/plugins/template/tests/functional/specific_properties_test.cpp
@@ -9,7 +9,7 @@
 #include "openvino/op/relu.hpp"
 #include "openvino/op/result.hpp"
 #include "openvino/runtime/properties.hpp"
-#include "openvino/util/common_util.hpp"
+#include "openvino/util/container_util.hpp"
 
 namespace ov::test {
