+++ src/tests/test_utils/common_test_utils/tests/utils_tests.cpp
@@ -7,6 +7,7 @@
 #include "openvino/util/common_util.hpp"
 #include "openvino/util/file_util.hpp"
 #include "openvino/util/env_util.hpp"
+#include "openvino/util/math_util.hpp"
 
 using namespace testing;
 using namespace ov::util;
